+++ client/App.tsx
@@ -1,8 +1,8 @@
 import React, { useEffect } from 'react';
-import { Redirect, Route, Switch } from "react-router";
+import { Route, Switch } from "react-router";
 import { toast, ToastContainer } from 'react-toastify';
 import { hot } from 'react-hot-loader';
-import { AnySSRPageData, RegenDBRequest } from "../server/routes/apiTypes";
+import { AnySSRPageData } from "../server/routes/apiTypes";
 import { usePageDataInit, PageDataContext } from "./hooks/usePageData";
 import { SSRProvider } from "./hooks/useSSR";
 import { ConfigContext } from "./hooks/useConfig";
@@ -10,13 +10,13 @@ import { CSRFContext } from "./hooks/useCSRF";
 import { ThemeProvider } from "./hooks/useTheme";
 import { PostsCacheProvider } from "./hooks/usePostsCache";
 import { QueryProvider } from "./hooks/useQuery";
-import requestJSON from "./helpers/requestJSON";
 import TestPage from "./routes/TestPage";
 import IndexPage from "./routes/index/IndexPage";
 import SearchPage from "./routes/search/SearchPage";
 import PostPage from "./routes/post/PostPage";
 import TagsPage from "./routes/tags/TagsPage";
 import NotFoundPage from "./routes/error/NotFoundPage";
+import DMCAPage from "./routes/dmca/DMCAPage";
 import "./globals.scss";
 
 interface Props {
@@ -63,6 +63,7 @@ export default hot(module)(function App({ initialData }: Props) {
                     <Route path="/tags" component={TagsPage} />
                     <Route path="/posts/:id" component={PostPage} />
                     <Route path="/posts" component={SearchPage} />
+                    <Route path="/dmca" component={DMCAPage} />
                     <Route path="/" exact component={IndexPage} />
                     <NotFoundPage />
                   </Switch>