+++ client/routes/dmca/DMCAPage.tsx
@@ -1,0 +1,7 @@
+import React from "react";
+import "./DMCAPage.scss";
+
+export default function DMCAPage() {
+  
+  return null;
+}